# Conflicts:
#	app/code/core/Mage/Catalog/Model/Factory.php
#	app/code/core/Mage/CatalogRule/Model/Action/Index/Refresh.php
#	tests/unit/Base/XmlFileLoadingTest.php
#	tests/unit/Mage/Admin/Model/UserTest.php
#	tests/unit/Mage/Core/Helper/EnvironmentConfigLoaderTest.php
#	tests/unit/Mage/Core/Helper/SecurityTest.php

+++ app/code/core/Mage/Catalog/Block/Product/Widget/Html/Pager.php
@@ -168,7 +168,7 @@ class Mage_Catalog_Block_Product_Widget_Html_Pager extends Mage_Page_Block_Html_
     public function getLastPageNum()
     {
         if ($this->_lastPage === null) {
-            $this->_lastPage = ceil($this->getCollectionSize() / $this->getLimit());
+            $this->_lastPage = (int) ceil($this->getCollectionSize() / $this->getLimit());
             if ($this->_lastPage <= 0) {
                 $this->_lastPage = 1;
             }
